requirements.txt
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Deprecated==1.2.18
Flask==3.1.1
frozenlist==1.6.2
future==1.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
line-bot-sdk==3.17.1
MarkupSafe==3.0.2
multidict==6.4.4
propcache==0.3.1
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.0
